refactor(step2): 스텝 메서드를 Station | Line | Path 기준으로 분리
