+++ README.md
@@ -60,3 +60,10 @@ cucumber가 제공하는 기능을 적절히 활용해보세요.
   - 인수 조건 정의 -> 인수 테스트 작성 -> 인수 테스트 충족 기능 구현 -> 리팩터링
   - 인수 조건을 주석으로 작성한다
   - 인수 테스트 이후 기능 구현은 TDD로 진행
+---
+## Feedback 24.03.12
+- [ ] `Path`는 경로 자체를 의미하니, 노선 목록 보다는 구간 목록이 자연스러워 보임
+- [ ] `Path`가 구간 목록을 상태로 가진다면, 객체의 응집도를 고려했을 때 요금 계산은 `Path`가 갖는 것이 자연스러워 보임
+- [ ] 요금 계산 하는 로직을 메서드를 분리해 볼 것
+- [x] 스텝 메서드의 배치와 분리를 고민해보기
+  - Station | Line | Path 기준으로 각각 분리
